many UI objects now pass down the arguments that __init__ gets passed to the parent this way certain things that inherit things like text can still be set no matter the object for example text for a button can be set because that gets passed to the "label" parent to display.
buttonsetcolor is now called highLightColor cause it's the color that indicates if a action has happened or is about to be taken on that button.

the base widget has a default size now and a id attribute
this id attribute can be used pass down some information regarding to the widget in question like for example it can be used to search for certain buttons with a certain id this can be handing if your program wants to only take a action on certain elements.

the parent attribute is added to widget it allows the widget to know the properties of its parent (a parent like a container) so that i can in theory use that information to for example subdivide the space or expand into the available space.

Idea: make a Container class that has a list of widgets and a add method both grid and box and pane will implement similar functionality to this and the main app class can then also be a container since it is a window that contains UI elements.

most if not all objects have a apply offset function inherited from Widget allowing them to shift their coordinates to the actual location so that drawing will take place properly and actions if possible can be taken properly on those elements.

the Widget attribute now also draws a empty box if there is no draw function implemented this allows for easy debuging.

moved the highlighting that i had implemented in CheckButton one level up into Button
because this is also something very useful in general button usage.

dubbing to rename it because highlighting something can also be when you hover over a button for example and it changes slightly of color.

maybe isHeld should be highLighted or something but then what is the Highlighting? maybe preselect? pre something else?
maybe something with indication as it could indicate what you are about to do.

the processEvents in Button has changed to only call releasedCallback if the mouse was on the button when it released and it has also changed the way the button indicates an action by allowing the button to return to the normal state even if the mouse moved off the button while the button was held down.

new Class Implemented named Indicator it can be used to Indicated a state it is set to I made it inherit from CheckButton but Button would have worked too? but it doesn't have to do any of the checks a button has to do it only has to visually show state whether that is for example some kind of button state or some other user defined state.

Have not made a lot of progress on the Grid Class but I don't know how I want to implement this I have decided to go for a simpler Object first namely boxes that can vertically or horizontally pack Widgets.