Renamed the function calcAlignment to apply because it applies the alignment parameters to the rect which can then be used to render what ever needs the alignment.
added two Classes that have to be implemented for Padding and Margin
Padding is for the inside of a widget for example between text and border in a label
while margin is for outside the widget for example between button and window border or another container border.

Widgets now have a update function the idea behind this is that it can update any element belonging to an widget so that
they appear in the right place.

Worked on the Container class and other Widgets that inherit from this
A container object will know about what widgets are inside.
and arrange them appropriately.

Working on the Box Container this will arrange widgets next to each other either vertically or horizontally.