The way update on widgets is propagated is changed now instead of setting the position and size for all the widgets first and then calling update on them all at once after is wrong they should be updated after setting those properties which makes sense otherwise they could go off of wrong values for their parent values.
the applyOffset() method has been removed the parent determines the assigned size it gets,
future have to figure out what to do with widgets when they get to small or if they have like a minimum size and such
but for now this is good enough.

the reason the parent sets the size on a widget through the setSize() function is that a widget can determine that that size is to small and maybe return something to indicate this so that for example containers could use this for determining widget location in the box and maybe even it's own size.