Redid alignment code it is easier to use and isn't a global instantiated object anymore. it doesn't need a padding or alignment parameter anymore those are now handled outside of alignment so that alignment is more of a general purpose aligning tool now.
rewrote all the widgets as far as I could find to use the new alignment/border/padding way of dividing up the space inside a widget.

colors for most widgets have changed a bit it isn't primary colors anymore but still testing things out.

buttons now change border color when you hover over them.

self.rect is no more it is now self.area (makes more sense widgets get an area assigned to them that they can use)

added MAGENTA to colorDefs

minor fix in container where draw would call self.drawBorder but that is called from draw() in Widget so no need to call it in your own draw function.

pane border now sorta green not sure about it yet not sure about the function of Pane's in general yet.

make box uses assigned content area instead of the old self.rect which would have incorrect divided up the widgets anyway.

minor issue with fontSizing where it was incorrectly assumed and making alignment slightly off for text in relation to the widget.

removed margin and padding .py files and classes realized they are the same thing and made a utility that can subtract two areas from each other to achieve generally what i needed for calculating padding alignment and border area.

debating whether alignment class really is just a thing that goes in rect utils it is pretty specific though and it has class attributes that are used to determine the type of alignment to apply not sure yet.

an attribute 'id' was added to Widget the idea was this could be used as an identifier to quickly find a widget in a container or something but debating about keeping it in.

Widget class had a major rewrite/addition in the shape of that it manages how to dividing space up inside it with margin/padding/border and content.
Three class attributes were added namely EXPAND, FILL and FIXED with FIXED the widget won't expand into the available space but instead will use a manually assigned area with EXPAND it will instead expand into the assigned space whether inside a container assigned space or in the main window assigned space.

FILL has to be implemented still but this simply is where padding is ignored and the widget used the full assigned area to display it's border and contents.

Widget also now has Methods for setting and getting it's position but they are just wrappers around the rectUtils set/get Area Position.
assignArea method has been modified to only apply a relative position if it's behaviour is set to FIXED.
a method for showing the layout was added to Widget this colors in the margin border padding and content area's to show their size position relative to the widget.