removed padding in buttons so that the border touches the content so that there isn't a visible gap making it not look alright.
default for containers to fill now (no margin area)

FILL in Widget implemented now.

started developing on a PackedBox but will leave this be for now, Wanted to do some cool things in combination with the logic simulation but ... the simulation code needs to be easier to work with
a better way to share input outputs names for such things basically a better interface to the code.
this will be a lot of work