Introduced Background drawing in Widget allowing you draw color for a background or override for specific backgrounds.
UI Callbacks need to change to be set to an object so that it can handle the details for calling a callback had to introduce
user arguments to make life a whole lot easier.

Made a PartBox this basically is a container that has elements that show the name of the part and it's inputs and outputs
and connects those inputs and outputs to buttons and indicators.