made the Instruction register clock in the instructions on the rising edge now one edge loads the instruction while the next can execute what ever that instruction has to do. I think this is how it works in the MC14500 but not 100% sure though. The result register has a "reset" sequence now in it's init.
Changed some more callbacks to the tuple way of (funcRef, args) type.
Checkbutton now has a function to uncheck.
Created a box in which you can click on instructions instead of entering the binary representation now you can just click a button and now what instruction will be put on the lines.